refactor(mail): give the search view its own header component
MailboxView had grown to 1905 lines, and the search query surface was the one
part of it that owed nothing to the rest: the field, the filter chips, and the
logic behind them (edit, toggle, remove, clear) were referenced from nowhere
else in the page, and every edit they make simply pushes a new route — which the
results below already read from.

So it moves out whole, needing no props. Only the search modal's three pieces of
state stay behind: the modal belongs to HeaderActions, and is opened from both
there and here, so it sits between them as models rather than local state.

MailboxView drops 136 lines for 39 added overall.

The obvious next targets don't pay for themselves. The shortcuts block would
need 25 bindings from the page and the action routers 13, so extracting either
buys a shorter file with a wider interface — more lines, and harder to follow.
What's left is one coupled state machine (resources, optimistic updates, undo,
focus, actions), and that coupling is real rather than incidental.
